Clean up folder structure
