+++ CleanCodeLesson.Tests/MagicWangTests.cs
@@ -1,4 +1,5 @@
 using CleanCodeLesson.Homework;
+using CleanCodeLesson.Homework.Cores;
 using FluentAssertions;
 
 namespace CleanCodeLesson.Tests;
