+++ CleanCodeLesson/Homework/Cores/ICore.cs
@@ -1,0 +1,6 @@
+namespace CleanCodeLesson.Homework.Cores;
+
+public interface ICore
+{
+    public Price Price { get; }
+}
