+++ CleanCodeLesson/Homework/Cores/PhoenixFeatherCore.cs
@@ -1,0 +1,6 @@
+namespace CleanCodeLesson.Homework.Cores;
+
+public record PhoenixFeatherCore : ICore
+{
+    public Price Price { get; } = new(4m);
+}
